Fixed page search filter:
Added a table to store the the title, standfirst, and bodycopy of the page for search indexes.

This simplifies the query for searches (removes the need for multiple joins on the chunk_texts table) and fixes an issue where pages where appearing in the search results multiple times if the search text was present in multiple fields.

Events are used to keep the search text table up to date. Only published (not draft nor embargoed) versions will appear in search (even for CMS users).