Removed duplication for chunk insert method
The chunk save controller and the page view controller both defined (inconcistent) functions for inserting a chunk

Also set the active page when saving a chunk.

These changes fix an issue when saving chunks which embed other chunks (e.g. when a page standfirst is used as part of a slideshow)