Removed the page withUrl() query scope
It was used when using the page finder to query pages.

It was put in to fix a problem which no longer seems to exist where pages where getting returned which couldn't be linked to.

However, it shouldn't be the default behaviour anyway since there may be times when pages don't have URLs but need to appear in a list of pages without links.

This should simplify the database queries for getting a list of pages.